+++ src/components/common/DataCreater.tsx
@@ -1,0 +1,531 @@
+import { FlightRow } from "./DataViewer";
+
+export const TempData : FlightRow[] = [{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+    pk: 1,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '부안',
+    frequency: 121.5,
+    txMain: '5/5',
+    rxMain: '5/5',
+    txStby: '5/5',
+    rxStby: '5/5',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+},
+{
+    site: '안양',
+    frequency: 121.4,
+    txMain: '5/4',
+    rxMain: '5/4',
+    txStby: '5/3',
+    rxStby: '5/6',
+    angle: 150,
+    distance: 35,
+    height: 5500,
+}   ]
