+++ src/components/common/DataPagination.tsx
@@ -1,0 +1,9 @@
+import { Pagination } from "@mui/material"
+
+const DataPagination = () => {
+    return (
+        <Pagination count={10} variant="outlined" shape={'rounded'}/>
+    )
+}
+
+export default DataPagination
